PoC: 모델을 float32로 통일 (fp16 체크포인트 입력 dtype 불일치 수정)

+++ train/poc_adapter.py
@@ -99,7 +99,9 @@ def main():
     device = "cuda" if torch.cuda.is_available() else "cpu"
     print(f"장치: {device} | 모델: {args.model} | 고스트 {args.layers}겹")
     proc = WhisperProcessor.from_pretrained(args.model)
-    model = WhisperForConditionalGeneration.from_pretrained(args.model).to(device)
+    # float32로 통일 — 체크포인트가 fp16이면 입력(float32)과 dtype이 안 맞아 에러난다.
+    # 어댑터 학습 안정성에도 fp32가 유리 (turbo 809M은 24GB GPU에 여유).
+    model = WhisperForConditionalGeneration.from_pretrained(args.model).to(device).float()
     model, trainable = attach_ghost(model, n_layers=args.layers)
     n_train_p = sum(p.numel() for p in trainable)
     print(f"학습 파라미터(고스트만): {n_train_p:,}")
